Correct dependencies for target dependencies in build graph.
Before, target dependencies depended only on their dependencies. This
is correct when paralellize builds is enabled, but it loses the order
when that's disabled. Correct it by storing the positional ordering
for those targets too.

Note this is kind of inefficient, we're representing a linear ordering
with vectors of increasing length.