Expose more values from the TargetEnvironment evaluation.
It's relatively expensive to do all of the computation in calculating
the TargetEnvironment. Currently, all of that computation is immediately
lost after the environment is discarded.

Two things are necesssary to preserve this calculation. First, the target
environment needs to expose more of its intermediate computed state so
it remains available later on. Second, and more importantly, the target
environments need to be cached (per BuildContext).

This implements the first (exposing intermediate state) portion of that.