Change plist copying and insertion to take std::unique_ptr.
This makes the semantics of creating or being passed an owned pointer
documented in the interface of the methods. Existing callsites are
updated to create or release the std::unique_ptr; future diffs will
make the value ownership explicit in the use cases as well.

A two-method approach is used to simulate covariance for copy().