Move Clang invocation creation to resolver model.
The CompilationInfo struct is slightly incomplete in that it holds
the full set of linker options, but the precompiled header info is
for a single compiler invocation, not the union of all compiler
invocations. This will be fixed by moving the precompiled header
creation logic into the compiler invocation context.