Have the Clang resolver add invocations directly to the tool context.
Rather than returning invocations from the tool resolvers, the new
method is to have them output into their tool context pointer. This
gives the resolvers more flexibilty to create multiple invocations
or zero invocations, as well as use the whole tool context for inter-
tool communication.

To begin, the Clang resolver is modified to take a tool context and
add its invocations directly to that. This also allows moving the
logic for precompiled header creation directly inside the Clang
resolver itself rather than having it be in the phase.