Support implicit dependencies through file references to a product.
Often, implicit dependencies are setup through a file reference into
the BUILT_PRODUCTS_DIR rather than a reference through a container
portal into another project. In those cases, match the product path
to the path referenced by the file reference. If they match, set up
an implicit dependency.

As an optimization, precompute the map of product path to target as
needed (depending on if implicit dependencies are allowed or not).