Use dynamic build rules for compile sources and copy resources phases.
These phases expect to be able to copy or compile any type of file, and
to use the target (and system) build rules to figure out how to. In order
to do that, refactor the resolution to use a centralized build rule
resolver.

Additionally, introduce the phase context as a phase-level counterpart
to the tool context, containing additionally cached tools for the build.