Implement Info.plist processing builtin tool.
The Info.plist processing tool is implemented, including conversions,
build setting expansion, and additional content files. Additionally, fix
a few bugs that caused the wrong set of arguments to be passed to it.

All builtin tools now take in an environment. This is used for Info.plist
processing to pass in the build settings to expand.