Simply ToolInvocation creation; separate out dependency types.
Rather than pass everything to the constructor, build up invocations
incrementally through mutable properties. Adjust all existing uses of
ToolInvocation to take a const value to preserve immutable semantics.

Dependencies can come in multiple types. Some are direct inputs to
the command. Others are input or output dependencies, in that they
are used by the command but are not the command's inputs or outputs.

Some inputs may be known to not always exist. Allow this by setting
them as "phony" inputs; the semantics here are the same as for Ninja.