Change formatter to print trailing, not leading, newlines.
This makes it easier to parse formatter return values because they
no longer start with a leading newline. Instead, print the newline
at the end of the previous section's output.

A number of steps in the formatter have added begin/finish variants
in order to support printing the trailing newlines at the right points.