Add very basic Ninja executor skeleton.
The Ninja executor can generate Ninja files for a subset of build
phases and invocations. It does not yet support builtin tools,
auxiliary files, or creating the output directory structure.

Additionally, extensions are added to the driver to support choosing
a formatter and executor dynamically. By default, the default
formatter and simple executor are still used. Pass `-executor ninja`
to use Ninja instead.