Add phony outputs, in addition to phony inputs.
Ninja can't handle duplicated outputs that go to the same path, so
use a phony output for the 'touch' rule. It will get its own phony
path so it doesn't conflict with the target to create the bundle
directory itself.

Additionally, phony inputs are separate from normal inputs; the same
value won't be in both. Update the simple executor to handle that.