Add executor support for creating product structure with invocations.
The simple executor now runs two phases of invocations, one for creating
the product structure, and one for running the main set of invocations.

The Formatter class gains an extra parameter indicating whether it should
format an invocation in a "simple" format, as used for the create product
structure invocations.

Invocations gain a flag if they are for creating the product structure.