Fix expansion of command lines with settings that have spaces inside.
If a tool's `CommandLine` had settings that expanded to have spaces,
before they would be incorrectly expanded into two arguments by splitting
on the space. For example:

    CommandLine = "clang $(InputPath)";

With the setting environment:

    InputPath = /path/with spaces/file.c
    OutputPath = /path/with spaces/file.o

Would turn into the argument list:

    ['clang', '/path/with', 'spaces/file.c']

Which is clearly incorrect. Now, the settings are expanded *after*
the string is split by token, which will end up with the correct result:

    ['clang', '/path/with spaces/file.c']