Avoid compiling header files with the default compiler.
Header files in copy resources phases have build rules applied. In
that case, must avoid processing the header with the default compiler
build rule. This is despite the fact that the build rule specifies
that all files of type `sourcecode.c`, which includes headers, should
be processed by the default compiler.

This is implemented by relying on the default compiler's limiting list
of file types. That limits the inputs to just actual source code files.