Make all non-required specification values optional.
This stores whether the values exist in the specification or are in
the specification but with an empty value. This is useful to have
correct behavior with regards to present but empty values.

Note inheritance is inverted with this diff: it's expected to be run
after the parse, not before. (That change is not included here.)