Fix duplicated results with build setting inheritance.
When build settings use $(inherited), previously it would duplicate
the result if there was any levels after where the setting is defined.
This was due to not incrementing the search iterator while resolving
assignment, so during inheritance it started at the beginning, not at
the current level.

Unit test added to prevent regressions.