Avoid copying build context or environment when resolving dependencies.
In general, avoiding copying large objects is good. However, this has
the side-effect of fixing a crash due to fragile use of the `project()`
method on `pbxproj::PBX::Target`. Targets only keep a weak reference to
their containing project, so the accesses are implicitly depending on
using the build context to keep the project alive. However, due to the
copy, only the transient build context would be kept alive.

Fixes #15.