Load nested projects at workspace creation time. Fixes #18.
When creating the workspace context, load all nested projects in the
workspace recursively. Then, collect all scheme groups for all those
projects, rather than just for the top-level projects.

It's unclear what environment to use to resolve the path to a file
reference for a nested project at workspace creation time. For now,
use the baseEnvironment and the project's top-level build settings.