Rename pbxbuild::Action to xcexecution, make dependencies explicit.
The action code depends on pbxbuild but nothing else in pbxbuild uses
it, so it's a good candidate for splitting out into a separate library.
Move it into a new xcexecution library that depends on pbxbuild, which is
now significantly smaller. This helps make dependencies explicit for
the code that, in pbxbuild, could have arbitrary circular dependencies.