Create representation of build parameters as the execution level.
In order for the executor to bypass loading the workspace when input
parameters have not changed, it needs a representation of the parameters
rather than having the workspace loaded before it starts.

A future change will pass xcexecution::Parameters into executors
instead of having them be created and used in the driver beforehand.