Automatically generate, re-generate, and run Ninja as appropriate.
The Ninja executor now automatically runs Ninja (or llbuild) as is
needed for build execution. This simplifies the Ninja executor usage
to a single command, rather than multiple commands and copy/paste.

In order to make this efficient, the Ninja is only regenerated if:

 - The build parameters change, based on the build parameter hash.
 - The inputs to the Ninja change. This is essentially files loaded
   in order to determine the build, like projects, schemes, and all
   specified xcconfig files.

A new extension option -generate is added to the driver in order to
accomodate regenerating from within an executor, which Ninja needs
to execute a command upon changes to any of its inputs.