Add ability to specify platforms and toolchains in a configuration file.
The optional configuration file is at `/var/db/xcsdk_configuration.plist`
unless `XCSDK_CONFIGURATION_PATH` is specified in the environment.

In the configuration, the key `ExtraPlatformsPaths` specifies paths with
additional `.platform` bundles to load, and `ExtraToolchainsPaths` lists
directories with additional `.xctoolchain` bundles.

A unit test is added to test that these paths are searched when specified.