Install xcrun aliases to $PATH.
In addition to installing all tools and libraries in
Developer/usr/, install xcrun in
${CMAKE_INSTALL_PREFIX}/bin (aka $PATH). Also install a
symlink pointing to xcrun called 'plutil' and make xcrun
automatically run plutil in this case.

The aliases in $PATH do not necessarily use xcbuild's
version of the tools. xcrun searches in ${DEVELOPER_DIR},
which may not point to xcbuild's Developer directory.