Write auxiliary files at build time, not Ninja generation time.
This allows Ninja to track dependencies properly for auxiliary files,
especially ones that are copied from source files rather than static.

Static auxiliary files are stored in the Ninja using Base64, since
some auxiliary files (e.g. header maps) are binary and thus are not
representable in shell scripts otherwise.

Explicit dependencies of invocations on their auxiliary files are
incomplete to that point that ordering between writing the auxiliary
files and the invocations is required; this is implemented with a
new barrier rule in Ninja between auxiliary files and invocations.