RFC Add reference classes to represent unloaded facets and renditions.
In many cases, fully parsing a facet or rendition is not needed. In
those cases, rather than expose the rendition data directly, expose
just a reference to the facet and rendition data within the archive
being read. The references can be used to realize a complete object
later, or passed directly to a writer to copy the data directly.

Simplify the fetch and iteration methods on the reader by switching
to directly referencing facets by identifier, rather than via name,
since the identifier is how the renditions are indexed. This allows
looking up renditions based on either a facet reference or complete
facet. Facets themselves are still indexed by name.

TODO: What if the key format for renditions doesn't match the format
from a reference? Need to account for that and adjust the renditions
as necessary to match, or adjust the key format itself.