[xcbuild] Check ~/.xcsdk/xcode_select_link when getting developer root
xcrun attempts to determine the developer root so that it knows where
how to locate toolchains and sdks.
Currently the developer root is resolved through the following steps
1. Checking DEVELOPER_DIR environment variable
2. Check the /var/db/xcode_select_link
3. /usr/share/xcode-select/xcode_dir_path
(1) suffers from having to configure this env var everytime xcrun is run.
(2) and (3) suffers on needing to be root to configure on most systems.

This change adds another symlink to check for:
~/.xcsdk/xcode_select_link

Test Plan:
1. Create symlink ~/.xcsdk/xcode_dir_path
2. xcrun -v -l --sdk test-sdk -f test-script

Change-Id: I88ed516e7264cffa263f6ed831e34083930e3337