[process] Don't use the constructor attribute on Linux
Using `__attribute__((constructor))` doesn't seem to get argc and argv
as arguments with certain versions of the libc. On the other hand, using
the `.init_array` section works so use that as a workaround until we can
find a cleaner fix.

h/t to @sas for this fix. This is a minimal fix for my problems when
using xcbuild tools on my Linux machine.

Fixes issue#138.