[process] Capture stdout/stderr in DefaultLauncher
== What's in this change? ==

* Modify the Launcher API such that it can return not only an exit code, but the
  stdout and stderr output as well.
* Modify DefaultLauncher such that it captures and returns stdout and
  stderr.

== Why make this change? ==

I'm working on adding an implementation of `swift-stdlib-tool` to
xcbuild. This tool is expected to, among other things, scan object files in
order to determine whether they contain references to the Swift
runtime. In order to implement this logic in xcbuild, I could write
C++ code that scans the object files, or I could use existing tools
that do so. It's my intention to shell out to these existing tools, but
to do so I'd need to:

1. Find the appropriate tool by running `xcrun -find` and capturing its
   output.
2. Running the tool found by step (1) and capturing its output. I'd then
  scan the output for text that indicates a Swift runtime dependency.

And so I need a utility in xcbuild that can capture output from
launched processes, which is what this change adds.

Change-Id: I0741dcdcbd6e7ab6202f6bec6447a36fb6d58334