libcar: silence a -Wsign-compare warning
When building libcar for Windows, the array size calculation would be of type
`unsigned long long` (aka `size_t`).  However, `identifier` is of type
`enum car_attribute_identifier` (aka `int`).  This results in a warning due to
the sign mismatch.  Perform an explicit cast to the size type for the
comparision to silence the warning.

Take the opportunity to move the constant expression of the array length into a
variable for ease of readability.